scripts/requirements.txt
annotated-types==0.7.0
arrow==1.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
flexcache==0.3
flexparser==0.4
idna==3.11
Pint==0.25
platformdirs==4.5.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
stravalib==2.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
